dndSpellsBackend/requirements.txt
asgiref==3.7.2
boto3==1.34.22
botocore==1.34.22
Django==5.0.1
django-filter==23.5
django-mongoengine-filter==0.4.2
django-storages==1.14.2
djangorestframework==3.14.0
dnspython==2.4.2
jmespath==1.0.1
Markdown==3.5.2
mongoengine==0.27.0
pillow==10.2.0
pymongo==4.6.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
s3transfer==0.10.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.4
urllib3==2.0.7
